+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/NoFilterEffect.java
@@ -1,0 +1,11 @@
+package com.otaliastudios.cameraview.filters;
+
+import com.otaliastudios.cameraview.filters.Filter;
+
+public class NoFilterEffect extends Filter {
+
+    @Override
+    public String getFragmentShader() {
+        return mFragmentShader;
+    }
+}
