+++ cameraview/src/main/java/com/otaliastudios/cameraview/preview/RendererFrameCallback.java
@@ -4,7 +4,7 @@ import android.graphics.SurfaceTexture;
 
 import androidx.annotation.NonNull;
 
-import com.otaliastudios.cameraview.shadereffects.BaseShaderEffect;
+import com.otaliastudios.cameraview.filters.Filter;
 
 /**
  * Callback for renderer frames.
@@ -30,5 +30,5 @@ public interface RendererFrameCallback {
      * @param scaleY the scaleY (in REF_VIEW) value
      */
     @RendererThread
-    void onRendererFrame(@NonNull SurfaceTexture surfaceTexture, float scaleX, float scaleY, BaseShaderEffect shaderEffect);
+    void onRendererFrame(@NonNull SurfaceTexture surfaceTexture, float scaleX, float scaleY, Filter shaderEffect);
 }
